pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-framework"
version = "0.1.0"
description = "Cohesive agentic framework with RAG, MCP, and self-correcting agents"
requires-python = ">=3.10"

dependencies = [
    # Core LLM and Agent Framework
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.50",
    "pydantic>=2.0.0",
    
    # RAG and Vector Search
    "langchain-text-splitters>=0.0.1",
    "embedchain>=0.2.0",
    "chromadb>=0.4.0",
    "pinecone-client>=3.0.0",
    
    # API Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    
    # MCP Server Support
    "mcp>=0.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "ipython>=8.15.0",
]

local-llm = [
    "ollama>=0.1.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
exclude = [".git", "__pycache__", "build", "dist"]
